share runWritableBeforeHooks logic
