+++ src/packages/core/src/base-resolver.ts
@@ -35,7 +35,7 @@ import { Sort, TypeMap } from './common/types';
 import { isExcludedFromFilterType, isReadOnlyBackend, isReadOnlyProperty } from './decorators';
 import { QueryManager } from './query-manager';
 import { HookManager, HookRegister } from './hook-manager';
-import { createOrUpdateEntities } from './utils/create-or-update-entities';
+import { createOrUpdateEntities, runWritableBeforeHooks } from './utils/create-or-update-entities';
 
 const arrayOperations = new Set(['in', 'nin']);
 const supportedOrderByTypes = new Set(['ID', 'String', 'Number', 'Date', 'ISOString']);
@@ -487,18 +487,6 @@ export function createBaseResolver<G extends WithId, D extends BaseDataEntity>(
 
 	@Resolver()
 	abstract class WritableBaseResolver extends BaseResolver {
-		private async runWritableBeforeHooks(
-			hookRegister: HookRegister,
-			params: CreateOrUpdateHookParams<G>
-		): Promise<CreateOrUpdateHookParams<G>> {
-			const hookManager = hookManagerMap.get(gqlEntityTypeName);
-			const hookParams = hookManager ? await hookManager.runHooks(hookRegister, params) : params;
-
-			const items = hookParams.args?.items;
-			if (!items) throw new Error('No data specified cannot continue.');
-			return params;
-		}
-
 		// Create many items in a transaction
 		@Mutation((returns) => [gqlEntityType], { name: `create${plural}` })
 		async createMany(
@@ -507,12 +495,16 @@ export function createBaseResolver<G extends WithId, D extends BaseDataEntity>(
 			@Ctx() context: BaseContext
 		): Promise<Array<G | null>> {
 			return this.withTransaction<Array<G | null>>(async () => {
-				const params = await this.runWritableBeforeHooks(HookRegister.BEFORE_CREATE, {
-					args: { items: createItems.data },
-					info,
-					context,
-					transactional,
-				});
+				const params = await runWritableBeforeHooks(
+					HookRegister.BEFORE_CREATE,
+					{
+						args: { items: createItems.data },
+						info,
+						context,
+						transactional,
+					},
+					gqlEntityTypeName
+				);
 				const { items } = params.args;
 				const entities = (await createOrUpdateEntities(
 					items,
@@ -532,12 +524,16 @@ export function createBaseResolver<G extends WithId, D extends BaseDataEntity>(
 			@Ctx() context: BaseContext
 		): Promise<G | null> {
 			return this.withTransaction<G | null>(async () => {
-				const params = await this.runWritableBeforeHooks(HookRegister.BEFORE_CREATE, {
-					args: { items: [createItemData] },
-					info,
-					context,
-					transactional,
-				});
+				const params = await runWritableBeforeHooks(
+					HookRegister.BEFORE_CREATE,
+					{
+						args: { items: [createItemData] },
+						info,
+						context,
+						transactional,
+					},
+					gqlEntityTypeName
+				);
 				const [item] = params.args.items;
 
 				const result = (await createOrUpdateEntities(item, gqlEntityType.name, info, context)) as G;
@@ -554,12 +550,16 @@ export function createBaseResolver<G extends WithId, D extends BaseDataEntity>(
 			@Ctx() context: BaseContext
 		): Promise<Array<G | null>> {
 			return this.withTransaction<Array<G | null>>(async () => {
-				const params = await this.runWritableBeforeHooks(HookRegister.BEFORE_UPDATE, {
-					args: { items: updateItems.data },
-					info,
-					context,
-					transactional,
-				});
+				const params = await runWritableBeforeHooks(
+					HookRegister.BEFORE_UPDATE,
+					{
+						args: { items: updateItems.data },
+						info,
+						context,
+						transactional,
+					},
+					gqlEntityTypeName
+				);
 				const { items } = params.args;
 
 				// Check that all objects have IDs
@@ -665,12 +665,16 @@ export function createBaseResolver<G extends WithId, D extends BaseDataEntity>(
 			@Ctx() context: BaseContext
 		): Promise<G | null> {
 			return this.withTransaction<G | null>(async () => {
-				const params = await this.runWritableBeforeHooks(HookRegister.BEFORE_UPDATE, {
-					args: { items: [updateItemData] },
-					info,
-					context,
-					transactional,
-				});
+				const params = await runWritableBeforeHooks(
+					HookRegister.BEFORE_UPDATE,
+					{
+						args: { items: [updateItemData] },
+						info,
+						context,
+						transactional,
+					},
+					gqlEntityTypeName
+				);
 				const [item] = params.args.items;
 
 				if (!updateItemData.id) throw new Error('No ID found in input so cannot update entity.');
